// venminder/vm-library/src/enums/vm-date-picker-css-definitions.enum.ts
export enum VMDatePickerCSSDefinitions {
    Selected = 'selected',
    Start = 'start',
    End = 'end',
    Lighter = 'lighter',
    DayNotInMonth = 'day-not-in-month',
    Today = 'today',
    FullView = 'full-view',
    BorderRight = 'border-right',
    RangeOptionsContainer = 'date-picker-container__rangeoptions',
    CalendarViewContainer = 'date-picker-container__calendar',
    DateRangeStyle = 'date-range',
}